spl_type_length_value::variable_len_pack

Trait VariableLenPack

pub trait VariableLenPack {
    // Required methods
    fn pack_into_slice(&self, dst: &mut [u8]) -> Result<(), ProgramError>;
    fn unpack_from_slice(src: &[u8]) -> Result<Self, ProgramError>
       where Self: Sized;
    fn get_packed_len(&self) -> Result<usize, ProgramError>;

    // Provided method
    fn pack(&self, dst: &mut [u8]) -> Result<(), ProgramError> { ... }
}

Trait that mimics a lot of the functionality of solana_program_pack::Pack but specifically works for variable-size types.

Required Methods

fn pack_into_slice(&self, dst: &mut [u8]) -> Result<(), ProgramError>

Writes the serialized form of the instance into the given slice

fn unpack_from_slice(src: &[u8]) -> Result<Self, ProgramError>
where Self: Sized,

Deserializes the type from the given slice

fn get_packed_len(&self) -> Result<usize, ProgramError>

Gets the packed length for a given instance of the type

Provided Methods

fn pack(&self, dst: &mut [u8]) -> Result<(), ProgramError>

Safely write the contents to the type into the given slice